Refactor Codebase and NodesStorage to use u32 for IDs
- Updated Codebase struct to replace fname_ast_map with syn_files and added files vector.
- Changed node ID type from u128 to u32 across various methods in Codebase and NodesStorage.
- Modified parse_and_add_file and build_api methods to accommodate new file handling.
- Adjusted tests to reflect changes in ID types and updated assertions accordingly.
- Ensured compatibility with new ID types in utility functions for mock data creation.

+++ sdk/src/ast/expression.rs
@@ -354,7 +354,7 @@ impl Binary {
     }
 
     #[must_use = "This method returns the id of the binary expression."]
-    pub fn id(&self) -> u128 {
+    pub fn id(&self) -> u32 {
         match self {
             Binary::Add(binex)
             | Binary::Sub(binex)
@@ -448,7 +448,7 @@ impl Unary {
     }
 
     #[must_use = "This method returns the id of the unary expression."]
-    pub fn id(&self) -> u128 {
+    pub fn id(&self) -> u32 {
         match self {
             Unary::Deref(unex) | Unary::Not(unex) | Unary::Neg(unex) => unex.id,
         }
